move sanitize_labels tests to test_utils
